+++ src/library/FileSearcher.h
@@ -13,7 +13,7 @@ public:
 		std::vector<std::string> _searchPaths;
 		std::vector<std::string> _excludePaths;
 		int _level { 0 };
-		int _size { 0 };
+		size_t _size { 0 };
 		std::vector<std::string> _masks;
 	};
